requirements.txt
# Core runtime dependencies
python-telegram-bot[job-queue]==20.7
python-dotenv==1.0.0
PyYAML==6.0.1
aiohttp>=3.9.0
psutil==5.9.8
colorama>=0.4.6
urllib3>=1.26.0
